Angular CheckAnswer service call.
One weird thing about this is that because of the way these
WebWorkClient scripts work, you can't interleave calls to any of them or
they'll just silently fail. This is because they basically login to a
fake class using the same account, and checkanswer.pl makes two requests
so if another request comes in between the two requests the session is
invalidated and the second request fails silently. Need to handle this
better.