Change how webwork_parser interprets division.
I don't know why it was handling division as a unary operator, but doing
it in the same way as other operations fixes the weird errors. I guess
the idea may have been to allow combining multiplications and divisons
more easily, not entirely sure. This fixes a lot of eval errors because
the add_header function did not expect this kind of operator.