Simple Python XMLRPC code which lets you render a PG file and check answers for it.
This will allow us to avoid the whole shelling out to a perl script song
and dance which should make the rendering and checking answer operations
an order of magnitude faster. Also, currently checkanswer relies on two
requests and if another call is made in between the session key becomes
invalidated and the second request fails. This avoids that whole
nonsense.